bot2/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "ML reinforcement learning bot for go-towerfall"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.7.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
    "websockets>=13.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bot"]

[tool.uv.sources]
torch = { index = "pytorch-cuda" }

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"

[tool.ruff]
src = ["src"]

[tool.ty.src]
exclude = ["examples/"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "ty>=0.0.4",
    "hatchling>=1.28.0",
]
